api/requirements.txt
peewee
requests
flask
unidecode
python-dateutil
flask-cors
python-dotenv
redis
pandas
